Make libpkg's checksum 64 bit compatible
Building The Heirloom Packaging Tools on GNU/Linux x86_64 produces wrong
checksums for pkgmap files. The issue is that they are calculated with
longs in compute_checksum(), which are expected to be 4 bytes long.
However, the x86_64 ABI spec makes long 8 bytes long.

In the illumos upstream repository this issue is fixed, so I changed
that one function to the version at
https://github.com/illumos/illumos-gate/blob/5c51f1241dbbdf2656d0e10011981411ed0c9673/usr/src/lib/libpkg/common/verify.c